+++ core/src/bricks/Snowpack.h
@@ -69,6 +69,11 @@ class Snowpack : public SurfaceComponent {
     void UpdateContentFromInputs() override;
 
     /**
+     * @copydoc Brick::ResetInputBooking()
+     */
+    void ResetInputBooking() override;
+
+    /**
      * @copydoc Brick::ApplyConstraints()
      */
     void ApplyConstraints(double timeStep) override;